safe reload of haproxies with consul template config in series (#278)

+++ ansible/roles/hcv-haproxy-configure/files/check-install-haproxy-config.sh
@@ -41,6 +41,11 @@ if [ $? -gt 0 ]; then
     FINAL_EXIT=1
 fi
 
+# skip id DRY_RUN is set
+if [ "$DRY_RUN" == "false" ]; then
+    echo "#### chic: in DRY_RUN mode" >> $TEMPLATE_LOGFILE
+fi
+
 if [ $FINAL_EXIT -eq 0 ]; then
     diff $DRAFT_CONFIG /etc/haproxy/haproxy.cfg
     if [ $? -gt 0 ]; then
@@ -48,21 +53,24 @@ if [ $FINAL_EXIT -eq 0 ]; then
 
         # save a copy of the new config
         cp "$DRAFT_CONFIG" $TEMPLATE_LOGDIR/$TIMESTAMP-haproxy.cfg
-
-        cp "$DRAFT_CONFIG" /etc/haproxy/haproxy.cfg
-        if [ $? -gt 0 ]; then
-            echo "#### chic: failed to copy the new haproxy config file to /etc/haproxy" >> $TEMPLATE_LOGFILE
-            FINAL_EXIT=1
-        else
-            service haproxy reload
+        if [ "$DRY_RUN" == "false" ]; then
+            cp "$DRAFT_CONFIG" /etc/haproxy/haproxy.cfg
             if [ $? -gt 0 ]; then
-                echo "#### chic: failed to reload haproxy service" >> $TEMPLATE_LOGFILE
+                echo "#### chic: failed to copy the new haproxy config file to /etc/haproxy" >> $TEMPLATE_LOGFILE
                 FINAL_EXIT=1
+            else
+                service haproxy reload
+                if [ $? -gt 0 ]; then
+                    echo "#### chic: failed to reload haproxy service" >> $TEMPLATE_LOGFILE
+                    FINAL_EXIT=1
+                else
+                    echo -n "jitsi.haproxy.reconfig:1|c" | nc -4u -w1 localhost 8125
+                    echo "#### chic: succeeded to reload haproxy with new config" >> $TEMPLATE_LOGFILE
+                fi
             fi
+        else
+            echo "#### cihc: validated $DRAFT_CONFIG; but DRY_RUN is not false, re-run with DRY_RUN=false to apply" >> $TEMPLATE_LOGFILE
         fi
-
-        echo -n "jitsi.haproxy.reconfig:1|c" | nc -4u -w1 localhost 8125
-        echo "#### chic: succeeded to reload haproxy with new config" >> $TEMPLATE_LOGFILE
     else 
         echo "#### cihc: validated $DRAFT_CONFIG; but new is the same as the old" >> $TEMPLATE_LOGFILE
     fi
